+++ src/Middleware/MiddlewareInterface.php
@@ -11,6 +11,7 @@ interface MiddlewareInterface
      * @param \Psr\Http\Message\ServerRequestInterface $request
      * @param \Psr\Http\Message\ResponseInterface $response
      * @param \WoohooLabs\Harmony\Harmony $next
+     * @return \Psr\Http\Message\ResponseInterface
      */
     public function __invoke(ServerRequestInterface $request, ResponseInterface $response, Harmony $next);
 }